pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zradci"
version = "0.1.0"
description = "Aplikace pro moderování hry Zrádci s email integrací"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "requests>=2.31.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "openai>=1.0.0",
    "email-validator>=2.3.0",
]

[project.scripts]
zradci = "main:app"

[tool.setuptools]
py-modules = ["main", "game_engine", "models", "email_sender", "config", "narrator", "email_receiver", "schemas", "voting"]

